labs/lab4-advanced-agents/pyproject.toml
[project]
name = "travel-concierge"
version = "0.1.0"
description = "Orchestrates personalized travel experiences, illustrates means to integrate booking systems, and provides support throughout the traveler's journey, from initial planning, booking, to real-time itinerary services and alerts."
authors = [
    { name = "Gary Ng", email = "gkcng@google.com" },
    { name = "Arun Santhanagopalan", email = "asanthan@google.com" },
    { name = "Renee Zhang", email = "xiaolizhang@google.com" },
    { name = "Anisha Kaul", email = "anishasjkaul@google.com" },
    { name = "Duncan Renfrow-Symon", email = "drenfrowsymon@google.com" },
]
license = "Apache-2.0"
readme = "README.md"
dependencies = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.126.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "google-genai>=1.16.1",
    "google-adk>=1.18.0",
    "google-cloud-storage>=3.0.0",
    "opentelemetry-instrumentation-google-genai",
    "opentelemetry-exporter-gcp-trace",
    "opentelemetry-exporter-gcp-logging",
    "mcp[cli]>=1.17.0",
]

requires-python = ">=3.10,<3.13"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "google-adk[eval]>=1.0.0",
    "pytest-asyncio>=0.26.0",
    "agent-starter-pack>=0.14.1",
]

deployment = [
    "absl-py>=2.2.1",
    "cloudpickle>=3.1.1",
    "flake8-pyproject>=1.2.3",
]

[project.optional-dependencies]
lint = [
    "ruff>=0.4.6",
    "mypy>=1.15.0",
    "codespell>=2.2.0",
    "types-pyyaml>=6.0.12.20240917",
    "types-requests>=2.32.0.20240914",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",  # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP", # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901"] # ignore line too long, too complex

[tool.ruff.lint.isort]
known-first-party = ["travel_concierge"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]

exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-root = ""

# This configuration file is used by goo.gle/agent-starter-pack to power remote templating.
# It defines the template's properties and settings.
[tool.agent-starter-pack]
example_question = "I want to plan a trip to Seattle"

[tool.agent-starter-pack.settings]
agent_directory = "travel_concierge"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true
